+++ src/main/java/linkedList/singly/JediLinkedList.java
@@ -1,4 +1,4 @@
-package linkedList;
+package linkedList.singly;
 
 public class JediLinkedList<T> {
