+++ src/main/java/linkedList/singly/JediLinkedListTest.java
@@ -1,4 +1,4 @@
-package linkedList;
+package linkedList.singly;
 
 public class JediLinkedListTest {
