+++ src/main/java/linkedList/singly/Node.java
@@ -1,4 +1,4 @@
-package linkedList;
+package linkedList.singly;
 
 public class Node<T> {
